refactor: rename ts folder to "app"
