+++ src/app/pages/service-pages/registration-page.ts
