+++ src/index.ts
@@ -1,4 +1,4 @@
-import { App } from './ts/app';
+import { App } from '@app/app';
 
 const app = new App();
 app.run();
